# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Dies ist eine OpenApi Spezifikation zur Implementierung eines Programmes, welches ToDo-Listen ausgibt, löscht, einfügt und bearbeitet. Das geschieht über verschiedene Endpunkte, die durch GET oder POST Requests angesprochen werden.",
    "version": "1.0.0",
    "title": "ToDo-API"
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/list/": {
      "post": {
        "tags": [
          "Liste"
        ],
        "description": "Erstellt eine neue Liste",
        "operationId": "addList",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "ListenInhalt",
            "description": "ToDo-Liste, die hinzugefügt werden soll.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/List"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/List"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/list/{list_id}": {
      "get": {
        "tags": [
          "Liste"
        ],
        "description": "Liefert alle Einträge einer Todo-Liste zurück.",
        "produces": [
          "application/json"
        ],
        "operationId": "getList",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "description": "ID der zu zeigenden Liste",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/List"
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Liste"
        ],
        "description": "Löscht eine Liste",
        "produces": [
          "application/json"
        ],
        "operationId": "deleteList",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "description": "ID der zu löschenden Liste",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/list/{list_id}/entry": {
      "post": {
        "tags": [
          "Eintrag"
        ],
        "description": "Fügt einen Eintrag zu einer bestehenden Todo-Liste hinzu.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "addEntry",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "description": "ID der zugehörigen Liste",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "Listeneintrag",
            "description": "Neuer Eintrag für eine bestehende Liste",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/List"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/list/{list_id}/entry/{entry_id}": {
      "post": {
        "tags": [
          "Eintrag"
        ],
        "description": "Aktualisiere einen Eintrag zu einer bestehenden Liste hinzu",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "updateEntry",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "description": "ID der zugehörigen Liste",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "entry_id",
            "description": "ID des Eintrags",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "neuerEintrag",
            "description": "Neues EintragObjekt",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Entry"
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Eintrag"
        ],
        "description": "Lösche einen Eintrag aus einer bestimmten Liste",
        "produces": [
          "application/json"
        ],
        "operationId": "deleteEntry",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "description": "ID der zugehörigen Liste",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "entry_id",
            "description": "ID des Eintrags",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "ZuLöschenderEintrag",
            "description": "Die ID des zu löschenden Eintrags sowie die ID der zugehörigen Liste",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Eine Liste aller Benutzer",
        "produces": [
          "application/json"
        ],
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/user/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Fügt einen neuen Benutzer hinzu",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "addUser",
        "parameters": [
          {
            "in": "body",
            "name": "NeuerBenutzer",
            "description": "JSON Objekt des neuen Nutzers",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/user/{user_id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Löscht einen bestehenden Nutzer",
        "produces": [
          "application/json"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "description": "Id des zu löschenden Nutzers",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
    "List": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "entries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Entry"
          }
        }
      }
    },
    "Entry": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "content": {
          "type": "string"
        },
        "listId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}